+++ SharpDPAPI/lib/Helpers.cs
@@ -126,6 +126,23 @@ namespace SharpDPAPI
 
             return HexAsBytes;
         }
+
+        public static string Capitalize(string s)
+        {
+            if(s.Length == 1)
+            {
+                return char.ToUpper(s[0]) + "";
+            }
+            else if (s.Length > 1)
+            {
+                return char.ToUpper(s[0]) + s.Substring(1);
+            }
+            else
+            {
+                return "";
+            }
+        }
+
         public static bool TestRemote(string computerName)
         {
             try
@@ -570,6 +587,11 @@ namespace SharpDPAPI
                    || array.Length == 0
                    || candidate.Length == 0
                    || candidate.Length > array.Length;
+        }
+
+        public static bool IsGuid(string value)
+        {
+            return Regex.IsMatch(value, @"^(\{{0,1}([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12}\}{0,1})$");
         }
     }
 }